Updated different files regarding paymentmethod
Updates files regarding payment method, changing it's type to Integer and the methods using it to fit accordingly

+++ backend/src/main/java/com/marcels/pos/services/SaleService.java
@@ -1,15 +1,15 @@
 package com.marcels.pos.services;
 
-import com.marcels.pos.models.entities.Sale;
-
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import com.marcels.pos.models.entities.Sale;
+
 public interface SaleService {
     List<Sale> getAllSales();
     List<Sale> getSalesByDateRange(LocalDateTime start, LocalDateTime end);
-    List<Sale> getSalesByPaymentMethod(Long paymentMethodId);
+    List<Sale> getSalesByPaymentMethod(Integer paymentMethodId);
     List<Sale> getSalesByMinimumAmount(BigDecimal minAmount);
     Sale getSaleById(Long id);
     Sale saveSale(Sale sale);
